MISC: Moved SetBitburnerDir to config/utils.go
