+++ utils/file.go
@@ -3,13 +3,8 @@ package utils
 import (
 	"fmt"
 	"os"
-	"path"
-	"runtime"
-	"strings"
 
 	"github.com/CTNOriginals/BitburnerGoFilesync/config"
-	"github.com/CTNOriginals/BitburnerGoFilesync/constants"
-
 	ctnfile "github.com/CTNOriginals/CTNGoUtils/v2/file"
 )
 
@@ -90,30 +85,4 @@ func SanitizeFileContent(content []byte) []byte {
 	}
 
 	return sanitized
-}
-
-func SetBitburnerDir(dir string) {
-	var isAbsolute = path.IsAbs(dir)
-
-	if runtime.GOOS == "windows" {
-		// checks if the second and third char of the path are ":/" or ":\"
-		isAbsolute = dir[1] == ':' && (dir[2] == '/' || dir[2] == '\\')
-	}
-
-	if !isAbsolute {
-		dir = fmt.Sprintf("%s/%s", constants.WorkindDirectory, dir)
-	}
-
-	// Replace all back slashes (\) with forward ones (/)
-	dir = strings.Join(strings.Split(dir, "\\"), "/")
-
-	if dir[len(dir)-1] == '/' {
-		dir = dir[0 : len(dir)-1]
-	}
-
-	dir = path.Clean(dir)
-
-	fmt.Printf("Set the bitburner working directory to: %s\n", dir)
-
-	config.Values.Directory = dir
 }
